Remove graphql vary and create cdnNormalizer.
