+++ src/service/graphql/utils/cacheControl.ts
@@ -67,18 +67,16 @@ export const cacheControl = (response: GraphQLResponse, ctx: GraphQLServiceConte
   const age = hints && minMaxAge(hints)
   const isPrivate = hints && anyPrivate(hints)
   const segment = hints && anySegment(hints)
-  const cacheControlHeaderScope = (isPrivate || isPrivateRoute(ctx)) ? 'private' : 'public'
-
-  ctx.graphql.cacheScope = segment ? 'segment' : cacheControlHeaderScope
 
   const maxAge = linked
     ? 'no-store'
     : (isPublicEndpoint(ctx) || !production)
       ? 'no-cache'
       : `max-age=${age}`
+
   return {
     maxAge,
-    scope: cacheControlHeaderScope,
+    scope: (isPrivate || isPrivateRoute(ctx)) ? 'private' : 'public',
     segment,
   }
 }
